+++ R/to-image.R
@@ -79,8 +79,9 @@ vw_to_svg <- function(spec, width = NULL, height = NULL, base_url = NULL,
   check_node_installed()
 
   # set defaults
-  base_url = base_url %||% ""
-  seed = seed %||% sample(1e8, size = 1)
+  base_url <-
+    base_url %||% getOption("vega.embed")[["loader"]][["baseURL"]] %||% ""
+  seed <- seed %||% sample(1e8, size = 1)
 
   spec <- vw_autosize(spec, width = width, height = height)
   vega_spec <- vw_to_vega(spec)